[BZ 1049054] remove legacy constructors from AggregateNumericMetric
There were a couple constructors that were used in conjuntion with the old
tables. Those have been removed along with some additional references to the
old tables. References to the old tables won't be completely removed though
until the metrics_cache and metrics_cache_index table are removed which will be
done under a separate BZ.