[BZ 1049054] log schedule ids for successfully migrated data
The migration records schedule ids of successful migrations in a log file that
is read at the start of the migration. If there are failures and the migration
has to be redo any work that has completed successfully.

This commit also updates MigrateAggregateMetricsTest.java so that it can be run
during jenkins builds. It now fully handles setting up and tearing down a
 single node cluster.